Fix packages on Object.prototype extension
Adding things to Object.prototype isn't great style, but users sometimes
do it.  Most of Meteor does fine with this because we mostly use
`_.each` and friends for loops, which skip over unowned properties. But
some pieces of Meteor don't do this.

In order to help us fix this, we add a method to Object.prototype
whenever you use tinytest.

XXX This commit does not completely work! `meteor test-packages` still
fails to work.  More changes like it are required.

Fixes #3478.